+++ Examples/Swift/Example-Swift-BridgingHeader.h
@@ -1,3 +1,1 @@
 #pragma once
-
-#import "MGLMapView+CustomAdditions.h"