+++ Examples/Swift/ViewController.swift
@@ -4,43 +4,25 @@ import MapboxNavigation
 import MapboxDirections
 import Mapbox
 
-let sourceIdentifier = "sourceIdentifier"
-let layerIdentifier = "layerIdentifier"
 private typealias RouteRequestSuccess = (([Route]) -> Void)
 private typealias RouteRequestFailure = ((NSError) -> Void)
 
-enum ExampleMode {
-    case `default`
-    case custom
-    case styled
-    case multipleWaypoints
-}
-
-class ViewController: UIViewController, MGLMapViewDelegate, CLLocationManagerDelegate {
-
-    // MARK: - Class Constants
-    static let mapInsets = UIEdgeInsets(top: 25, left: 25, bottom: 25, right: 25)
+class ViewController: UIViewController, MGLMapViewDelegate {
     
     // MARK: - IBOutlets
     @IBOutlet weak var longPressHintView: UIView!
-
     @IBOutlet weak var simulationButton: UIButton!
     @IBOutlet weak var startButton: UIButton!
     @IBOutlet weak var bottomBar: UIView!
-    
     @IBOutlet weak var clearMap: UIButton!
 
     // MARK: Properties
     var mapView: NavigationMapView?
-    var waypoints: [Waypoint] = []
-    var currentRoute: Route? {
-        get {
-            return routes?.first
-        }
-        set {
-            guard let selected = newValue else { routes?.remove(at: 0); return }
-            guard let routes = routes else { self.routes = [selected]; return }
-            self.routes = [selected] + routes.filter { $0 != selected }
+    var waypoints: [Waypoint] = [] {
+        didSet {
+            waypoints.forEach {
+                $0.coordinateAccuracy = -1
+            }
         }
     }
 
@@ -62,6 +44,8 @@ class ViewController: UIViewController, MGLMapViewDelegate, CLLocationManagerDel
         self?.mapView?.removeWaypoints()
         self?.routes = routes
         self?.waypoints = current.routeOptions.waypoints
+        self?.clearMap.isHidden = false
+        self?.longPressHintView.isHidden = true
     }
 
     fileprivate lazy var defaultFailure: RouteRequestFailure = { [weak self] (error) in
@@ -69,49 +53,37 @@ class ViewController: UIViewController, MGLMapViewDelegate, CLLocationManagerDel
         print(error.localizedDescription)
     }
 
-    var exampleMode: ExampleMode?
-
-    var locationManager = CLLocationManager()
-
     var alertController: UIAlertController!
-
-    lazy var multipleStopsAction: UIAlertAction = {
-        return UIAlertAction(title: "Multiple Stops", style: .default, handler: { (action) in
-            self.startMultipleWaypoints()
-        })
-    }()
 
     // MARK: - Lifecycle Methods
 
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        locationManager.delegate = self
-        locationManager.requestWhenInUseAuthorization()
-
-        automaticallyAdjustsScrollViewInsets = false
-
-        simulationButton.isSelected = true
-        startButton.isEnabled = false
+        CLLocationManager().requestWhenInUseAuthorization()
         
         alertController = UIAlertController(title: "Start Navigation", message: "Select the navigation type", preferredStyle: .actionSheet)
-        alertController.addAction(UIAlertAction(title: "Default UI", style: .default, handler: { (action) in
-            self.startBasicNavigation()
-        }))
-        alertController.addAction(UIAlertAction(title: "DayStyle UI", style: .default, handler: { (action) in
-            self.startNavigation(styles: [DayStyle()])
-        }))
-        alertController.addAction(UIAlertAction(title: "NightStyle UI", style: .default, handler: { (action) in
-            self.startNavigation(styles: [NightStyle()])
-        }))
-        alertController.addAction(UIAlertAction(title: "Custom UI", style: .default, handler: { (action) in
-            self.startCustomNavigation()
-        }))
-        alertController.addAction(UIAlertAction(title: "Styled UI", style: .default, handler: { (action) in
-            self.startStyledNavigation()
-        }))
-
-        alertController.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: nil))
+        
+        typealias ActionHandler = (UIAlertAction) -> Void
+        
+        let basic: ActionHandler = {_ in self.startBasicNavigation() }
+        let day: ActionHandler = {_ in self.startNavigation(styles: [DayStyle()]) }
+        let night: ActionHandler = {_ in self.startNavigation(styles: [NightStyle()]) }
+        let custom: ActionHandler = {_ in self.startCustomNavigation() }
+        let styled: ActionHandler = {_ in self.startStyledNavigation() }
+        
+        let actionPayloads: [(String, UIAlertActionStyle, ActionHandler?)] = [
+            ("Default UI", .default, basic),
+            ("DayStyle UI", .default, day),
+            ("NightStyle UI", .default, night),
+            ("Custom UI", .default, custom),
+            ("Styled UI", .default, styled),
+            ("Cancel", .cancel, nil)
+        ]
+        
+        actionPayloads
+            .map { payload in UIAlertAction(title: payload.0, style: payload.1, handler: payload.2)}
+            .forEach(alertController.addAction(_:))
 
         if let popoverController = alertController.popoverPresentationController {
             popoverController.sourceView = self.startButton
@@ -128,20 +100,12 @@ class ViewController: UIViewController, MGLMapViewDelegate, CLLocationManagerDel
         if (presentedViewController != nil) {
             DayStyle().apply()
         }
-        
     }
 
     // MARK: Gesture Recognizer Handlers
 
     @objc func didLongPress(tap: UILongPressGestureRecognizer) {
-        guard tap.state == .began else {
-            return
-        }
-
-        clearMap.isHidden = false
-        longPressHintView.isHidden = true
-        
-        guard let mapView = mapView else { return }
+        guard let mapView = mapView, tap.state == .began else { return }
 
         if let annotation = mapView.annotations?.last, waypoints.count > 2 {
             mapView.removeAnnotation(annotation)
@@ -149,22 +113,12 @@ class ViewController: UIViewController, MGLMapViewDelegate, CLLocationManagerDel
 
         if waypoints.count > 1 {
             waypoints = Array(waypoints.suffix(1))
-            multipleStopsAction.isEnabled = true
-        } else { //single waypoint
-            multipleStopsAction.isEnabled = false
         }
-
         
         let coordinates = mapView.convert(tap.location(in: mapView), toCoordinateFrom: mapView)
         // Note: The destination name can be modified. The value is used in the top banner when arriving at a destination.
-        let destinationName = "Dropped Pin #\(waypoints.endIndex + 1)"
-        let waypoint = Waypoint(coordinate: coordinates, name: destinationName)
-        waypoint.coordinateAccuracy = -1
+        let waypoint = Waypoint(coordinate: coordinates, name: "Dropped Pin #\(waypoints.endIndex + 1)")
         waypoints.append(waypoint)
-
-        if waypoints.count >= 2, !alertController.actions.contains(multipleStopsAction) {
-            alertController.addAction(multipleStopsAction)
-        }
 
         requestRoute()
     }
@@ -192,7 +146,7 @@ class ViewController: UIViewController, MGLMapViewDelegate, CLLocationManagerDel
         mapView?.removeRoutes()
         mapView?.removeWaypoints()
         waypoints.removeAll()
-        multipleStopsAction.isEnabled = false
+        longPressHintView.isHidden = false
     }
 
     @IBAction func startButtonPressed(_ sender: Any) {
@@ -205,8 +159,7 @@ class ViewController: UIViewController, MGLMapViewDelegate, CLLocationManagerDel
         guard waypoints.count > 0 else { return }
         guard let mapView = mapView else { return }
 
-        let departureName = "User location"
-        let userWaypoint = Waypoint(location: mapView.userLocation!.location!, heading: mapView.userLocation?.heading, name: departureName)
+        let userWaypoint = Waypoint(location: mapView.userLocation!.location!, heading: mapView.userLocation?.heading, name: "User location")
         waypoints.insert(userWaypoint, at: 0)
 
         let options = NavigationRouteOptions(waypoints: waypoints)
@@ -228,9 +181,7 @@ class ViewController: UIViewController, MGLMapViewDelegate, CLLocationManagerDel
     // MARK: Basic Navigation
 
     func startBasicNavigation() {
-        guard let route = currentRoute else { return }
-
-        exampleMode = .default
+        guard let route = routes?.first else { return }
 
         let navigationViewController = NavigationViewController(for: route, locationManager: navigationLocationManager())
         navigationViewController.delegate = self
@@ -239,24 +190,19 @@ class ViewController: UIViewController, MGLMapViewDelegate, CLLocationManagerDel
     }
     
     func startNavigation(styles: [Style]) {
-        guard let route = currentRoute else { return }
-        
-        exampleMode = .default
+        guard let route = routes?.first else { return }
         
         let navigationViewController = NavigationViewController(for: route, styles: styles, locationManager: navigationLocationManager())
         navigationViewController.delegate = self
         
         presentAndRemoveMapview(navigationViewController)
     }
-
+    
     // MARK: Custom Navigation UI
-
     func startCustomNavigation() {
-        guard let route = self.currentRoute else { return }
+        guard let route = routes?.first else { return }
 
         guard let customViewController = storyboard?.instantiateViewController(withIdentifier: "custom") as? CustomViewController else { return }
-
-        exampleMode = .custom
 
         customViewController.userRoute = route
 
@@ -271,9 +217,7 @@ class ViewController: UIViewController, MGLMapViewDelegate, CLLocationManagerDel
     // MARK: Styling the default UI
 
     func startStyledNavigation() {
-        guard let route = self.currentRoute else { return }
-
-        exampleMode = .styled
+        guard let route = routes?.first else { return }
 
         let styles = [CustomDayStyle(), CustomNightStyle()]
 
@@ -284,23 +228,10 @@ class ViewController: UIViewController, MGLMapViewDelegate, CLLocationManagerDel
     }
 
     func navigationLocationManager() -> NavigationLocationManager {
-        guard let route = currentRoute else { return NavigationLocationManager() }
+        guard let route = routes?.first else { return NavigationLocationManager() }
         return simulationButton.isSelected ? SimulatedLocationManager(route: route) : NavigationLocationManager()
     }
 
-    // MARK: Navigation with multiple waypoints
-
-    func startMultipleWaypoints() {
-        guard let route = self.currentRoute else { return }
-
-        exampleMode = .multipleWaypoints
-
-        let navigationViewController = NavigationViewController(for: route, locationManager: navigationLocationManager())
-        navigationViewController.delegate = self
-
-        presentAndRemoveMapview(navigationViewController)
-    }
-    
     func presentAndRemoveMapview(_ navigationViewController: NavigationViewController) {
         present(navigationViewController, animated: true) {
             self.mapView?.removeFromSuperview()
@@ -317,33 +248,21 @@ class ViewController: UIViewController, MGLMapViewDelegate, CLLocationManagerDel
         mapView.delegate = self
         mapView.navigationMapDelegate = self
         mapView.userTrackingMode = .follow
-        let bottomPadding = (view.frame.height + view.frame.origin.y) - bottomBar.frame.origin.y
         
-        var topPadding: CGFloat = 0.0
-        if #available(iOS 11.0, *) {
-            topPadding = view.safeAreaInsets.top
-        } else if let navCon = navigationController {
-            topPadding = navCon.navigationBar.frame.size.height
-        }
-        
-        let subviewMask = UIEdgeInsets(top: topPadding, left: 0, bottom: bottomPadding, right: 0)
-        mapView.contentInset = ViewController.mapInsets + subviewMask
-        
-        view.insertSubview(mapView, belowSubview: bottomBar)
+        view.insertSubview(mapView, belowSubview: longPressHintView)
         
         let singleTap = UILongPressGestureRecognizer(target: self, action: #selector(didLongPress(tap:)))
         mapView.gestureRecognizers?.filter({ $0 is UILongPressGestureRecognizer }).forEach(singleTap.require(toFail:))
         mapView.addGestureRecognizer(singleTap)
-    
     }
     
     func mapView(_ mapView: MGLMapView, didFinishLoading style: MGLStyle) {
         self.mapView?.localizeLabels()
         
-        if let routes = routes, let coords = currentRoute?.coordinates, let coordCounts = currentRoute?.coordinateCount {
-            mapView.setVisibleCoordinateBounds(MGLPolygon(coordinates: coords, count: coordCounts).overlayBounds, animated: false)
+        if let routes = routes, let currentRoute = routes.first, let coords = currentRoute.coordinates {
+            mapView.setVisibleCoordinateBounds(MGLPolygon(coordinates: coords, count: currentRoute.coordinateCount).overlayBounds, animated: false)
             self.mapView?.showRoutes(routes)
-            self.mapView?.showWaypoints(currentRoute!)
+            self.mapView?.showWaypoints(currentRoute)
         }
     }
 }
@@ -351,7 +270,7 @@ class ViewController: UIViewController, MGLMapViewDelegate, CLLocationManagerDel
 // MARK: - NavigationMapViewDelegate
 extension ViewController: NavigationMapViewDelegate {
     func navigationMapView(_ mapView: NavigationMapView, didSelect waypoint: Waypoint) {
-        guard let routeOptions = currentRoute?.routeOptions else { return }
+        guard let routeOptions = routes?.first?.routeOptions else { return }
         let modifiedOptions = routeOptions.without(waypoint: waypoint)
 
         presentWaypointRemovalActionSheet { _ in
@@ -360,7 +279,10 @@ extension ViewController: NavigationMapViewDelegate {
     }
 
     func navigationMapView(_ mapView: NavigationMapView, didSelect route: Route) {
-        currentRoute = route
+        guard let routes = routes else { return }
+        guard let index = routes.index(where: { $0 == route }) else { return }
+        self.routes!.remove(at: index)
+        self.routes!.insert(route, at: 0)
     }
 
     private func presentWaypointRemovalActionSheet(completionHandler approve: @escaping ((UIAlertAction) -> Void)) {
@@ -422,9 +344,6 @@ extension ViewController: NavigationViewControllerDelegate {
     // Return false to remain on the current leg, for example to allow the user to provide input.
     // If you return false, you must manually advance to the next leg. See the example above in `confirmationControllerDidConfirm(_:)`.
     func navigationViewController(_ navigationViewController: NavigationViewController, didArriveAt waypoint: Waypoint) -> Bool {
-        // Multiple waypoint demo
-        guard exampleMode == .multipleWaypoints else { return true }
-
         // When the user arrives, present a view controller that prompts the user to continue to their next destination
         // This type of screen could show information about a destination, pickup/dropoff confirmation, instructions upon arrival, etc.
         guard let confirmationController = self.storyboard?.instantiateViewController(withIdentifier: "waypointConfirmation") as? WaypointConfirmationViewController else { return true }
@@ -442,35 +361,16 @@ extension ViewController: NavigationViewControllerDelegate {
     }
 }
 
-/**
- To find more pieces of the UI to customize, checkout DayStyle.swift.
- */
-// MARK: CustomDayStyle
-class CustomDayStyle: DayStyle {
-    
-    required init() {
-        super.init()
-        mapStyleURL = URL(string: "mapbox://styles/mapbox/satellite-streets-v9")!
-        styleType = .day
-    }
-    
-    override func apply() {
-        super.apply()
-        BottomBannerView.appearance().backgroundColor = .orange
-    }
-}
-
-// MARK: CustomNightStyle
-class CustomNightStyle: NightStyle {
-
-    required init() {
-        super.init()
-        mapStyleURL = URL(string: "mapbox://styles/mapbox/satellite-streets-v9")!
-        styleType = .night
-    }
-
-    override func apply() {
-        super.apply()
-        BottomBannerView.appearance().backgroundColor = .purple
+// Mark: VisualInstructionDelegate
+extension ViewController: VisualInstructionDelegate {
+    func label(_ label: InstructionLabel, willPresent instruction: VisualInstruction, as presented: NSAttributedString) -> NSAttributedString? {
+        
+        // Uncomment to mutate the instruction shown in the top instruction banner
+        // let range = NSRange(location: 0, length: presented.length)
+        // let mutable = NSMutableAttributedString(attributedString: presented)
+        // mutable.mutableString.applyTransform(.latinToKatakana, reverse: false, range: range, updatedRange: nil)
+        // return mutable
+        
+        return presented
     }
 }